Run commands in their own process group id
This will avoid the terminal sending the interrupt
signal to the parent and child processes since
their are not in the same process group id.

Before this, if you press control+c, the signal will
be received by the `main.go` program, and all its
running `commands`, now, it's only received by the
`main.go` program.