Restore service status after updating
If there is an error while updating, the service's status is set to
`command.StatusRunning` and this cannot be true. If the process is
stopped, and after updating the function fail, is going to set the
status to `command.StatusRunning` and this is wrong. Now we store the
status before mutate it and we restore it if either updater fails or
if is not updated.